error

+++ Vehicle.Application/Features/Customers/Commands/DeleteCustomerCommandHandler.cs
@@ -6,8 +6,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Vehicle.Application.Common.Exceptions;
 using Vehicle.Application.Contracts.Persistence;
 using Vehicle.Domain.Entities;
+using Vehicle.Domain.Entities.Concrete;
 
 namespace Vehicle.Application.Features.Customers.Commands
 {
@@ -27,6 +29,10 @@ namespace Vehicle.Application.Features.Customers.Commands
         public async Task<Unit> Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
         {
             var customerToDelete = await _customerRepository.GetByIdAsync(request.Id);
+            if (customerToDelete == null)
+            {
+                throw new NotFoundException(nameof(Customer), request.Id);
+            }
 
             await _customerRepository.DeleteAsync(customerToDelete);
             _logger.LogInformation($"the {customerToDelete.Id} is successfully deleted.");
